Cast the expression tp-tuples to int, this is needed on platforms where sizeof(int) != sizeof(ptrdiff_t), such as x86_64.

+++ src/perpet.c
@@ -974,9 +974,9 @@ int main(int argc, char *argv[])
 		   {
 		     if(!tp->abstainable) continue;
 		     if(tp->exechance > 0)
-		       (void) fprintf(ofp, "ICK_INT1 ICKABSTAINED(%d)=0;\n",tp-tuples);
+		       (void) fprintf(ofp, "ICK_INT1 ICKABSTAINED(%d)=0;\n",(int)(tp-tuples));
 		     else
-		       (void) fprintf(ofp, "ICK_INT1 ICKABSTAINED(%d)=1;\n",tp-tuples);
+		       (void) fprintf(ofp, "ICK_INT1 ICKABSTAINED(%d)=1;\n",(int)(tp-tuples));
 		   }
 		 }
 		 if(!pickcompile) (void) fprintf(ofp, "}");
